fix: Fix/re order internal accounts (#9993)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

<!--
Write a short description of the changes included in this pull request,
also include relevant motivation and context. Have in mind the following
questions:
1. What is the reason for the change?
2. What is the improvement/solution?
-->

This PR updates the `selectInternalAccounts` selector to return accounts
sorted by the order of the accounts in the KeyringController. This
update preserves the sorting behavior of accounts and matches behavior
of versions prior to 7.24.0.

## **Related issues**

Fixes: https://github.com/MetaMask/metamask-mobile/issues/9990

## **Manual testing steps**

To reproduce the issue

- Install 7.23.0
- Import accounts via SRP
- Create an account from the wallet
- Import an account via private key
- Create another account from the wallet
- Notice that the section of accounts on the account selector list is
ordered by 1. SRP, 2.. Wallet, 3. Private key

- Download 7.24.0
- Open the account selector list and notice that accounts are now sorted
by the order in which the accounts were created. Reference the steps
above for the order.

To see the fix
- Install this branch
- Open the account selector list and notice that the accounts are
ordered the same as it was on 7.23.0 - 1. SRP, 2.. Wallet, 3. Private
key.

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->
Creating accounts on 7.21.0, then upgrading to 7.24.0. The account name
change is unrelated to this fix and will be addressed in a separate
hotfox for 7.24.1.

https://github.com/MetaMask/metamask-mobile/assets/10508597/3c1a5587-1172-4db1-8bea-6d64239f7baf

### **After**

<!-- [screenshots/recordings] -->
Continuing from the state above, upgrading to this branch fixes the
ordering

https://github.com/MetaMask/metamask-mobile/assets/10508597/15277032-2adf-4370-bd3c-c92b3037c526


## **Pre-merge author checklist**

- [x] I’ve followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask Mobile
Coding
Standards](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [x] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [x] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.