feat: Add team label to pr (#10086)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

<!--
Write a short description of the changes included in this pull request,
also include relevant motivation and context. Have in mind the following
questions:
1. What is the reason for the change?
2. What is the improvement/solution?
-->

This PR adds the `add-team-label` GitHub workflow and the accompanying
`add-team-label-to-pr` script. Most of the implementation follows the
`add-release-label` GitHub workflow and the
`add-release-label-to-pr-and-linked-issues` script. To make the
implementation easier, a new helper function, `addLabelByIdToLabelable`
was also added, and the previously non-exported `retrieveLabel` function
was exported.

When a new PR is opened, it will automatically add the author's team to
the labels.

## **Related issues**

Fixes: https://github.com/MetaMask/MetaMask-planning/issues/2447

## **Manual testing steps**

1. Open PR, and see that the correct label is added.

## **Screenshots/Recordings**

Not applicable

## **Pre-merge author checklist**

- [x] I’ve followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask Mobile
Coding
Standards](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.