test: Fix Swap e2e test failure (#10897)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

This fixes the test failures of swap e2e regression tests which are
caused by the following problem.

When swapping two tokens a quote is displayed for about 30 seconds.
After the quote is displayed on Android it can now take about 15 seconds
before the app becomes idle. But after 15 seconds the gas fee start to
flash and new fees are displayed so during that time the swap button
becomes disabled for a few seconds. Because of the prolong wait due to
the app not becoming idle the script was trying to swap when the button
was disabled and fail. So to be safe I have changed the script to wait
for 20 seconds before swiping.

I did 3 bitrise runs and all passed

https://app.bitrise.io/app/be69d4368ee7e86d/pipelines/ca7de740-a15e-4bf4-af87-71f2bad20e8c

https://app.bitrise.io/app/be69d4368ee7e86d/pipelines/b4a56a55-d520-4d57-a1eb-2307001bf55c

https://app.bitrise.io/app/be69d4368ee7e86d/pipelines/9e73ba7b-c3ca-4010-87ab-f8e239be0652

## **Related issues**

Fixes:

## **Manual testing steps**

1. Go to this page...
2.
3.

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->

### **After**

<!-- [screenshots/recordings] -->

## **Pre-merge author checklist**

- [x] I’ve followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask Mobile
Coding
Standards](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [x] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [x] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.